实现第三方接口整合

+++ heima-leadnews-model/src/main/java/com/heima/model/media/pojos/WmUser.java
@@ -26,13 +26,13 @@ public class WmUser implements Serializable {
      * 主键
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     @TableField("ap_user_id")
-    private Integer apUserId;
+    private Long apUserId;
 
     @TableField("ap_author_id")
-    private Integer apAuthorId;
+    private Long apAuthorId;
 
     /**
      * 登录用户名
